requirements.txt
requests
azure-identity 
azure-keyvault-secrets
azure-storage-blob
